// TypeScript/src/description.enum.ts
export enum Description {
    GOOD_WINE = 'Good Wine',
    BACKSTAGE_PASSES_FOR_RE_FACTOR = 'Backstage passes for Re:Factor',
    BACKSTAGE_PASSES_FOR_HAXX = 'Backstage passes for HAXX',
    B_DAWG_KEYCHAIN = 'B-DAWG Keychain',
    RING_OF_CLEANSENING_CODE = 'Ring of Cleansening Code',
    DUPLICATE_CODE = 'Duplicate Code',
    LONG_METHODS = 'Long Methods',
    UGLY_VARIABLE_NAMES = 'Ugly Variable Names'
}